feat(fiber_network): handle explicit instantiation; clean up; fix test
Main changes:
 - With explicit instantition, we do not need to give the path to source file when using the FiberNetwork class;
 - Add assertions;
 - Avoid using ``new";
 - Output ArrayView in get_fibers function;
 - Delete destructor;
 - Delete unimportant header files
