+++ tests/mechanics/fiber_network_01.input
@@ -1,0 +1,3 @@
+Main {
+    log_file_name = "fiber_output"
+}
