+++ tests/mechanics/fiber_network_01.output
@@ -1,3 +1,4 @@
+Test with two constant vector fields
 0 1 0 0 1
 1 1 0 0 1
 2 1 0 0 1
